Moved ProtocolVersion code to separate file
As this code is something we want to keep while reimplementing
parts of the network protocol, I initially thought to place it
in protocol.h.

However, since it seems like something I'll want to use from
both new and old code having it in a separate file makes it
easier to include, since protocol.h already depends on cfnet.h
Making cfnet.h depend on protocol.h would create a circular
include. The solution is to create a completely separate file
protocol_version, which #includes (almost) nothing, so both
cfnet.h and protocol.h can include it.

Changelog: None
Ticket: None
